requirements.txt
asgiref==3.6.0
attrs==22.2.0
charset-normalizer==3.1.0
colorama==0.4.6
crispy-bootstrap4==2022.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.7
django-crispy-forms==2.0
django-import-export==3.1.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.9.0
idna==3.4
iniconfig==2.0.0
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
phonenumbers==8.13.7
Pillow==9.5.0
platformdirs==3.1.1
pluggy==1.0.0
pytest==7.2.2
PyYAML==6.0
sqlparse==0.4.3
tablib==3.4.0
tomli==2.0.1
tzdata==2022.7
urllib3==2.0.2
virtualenv==20.20.0
xlrd==2.0.1
xlwt==1.3.0
